+++ src/main/java/com/forgestove/create_cyber_goggles/CreateCyberGoggles.java
@@ -1,21 +1,41 @@
 package com.forgestove.create_cyber_goggles;
-import com.forgestove.create_cyber_goggles.content.ModConfig;
-import me.shedaniel.autoconfig.AutoConfig;
-import me.shedaniel.autoconfig.serializer.Toml4jConfigSerializer;
+import com.forgestove.create_cyber_goggles.content.config.CyberConfig;
+import com.forgestove.create_cyber_goggles.content.event.*;
 import net.neoforged.api.distmarker.Dist;
+import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.ModContainer;
-import net.neoforged.fml.common.Mod;
-import net.neoforged.neoforge.client.gui.IConfigScreenFactory;
+import net.neoforged.fml.common.*;
+import net.neoforged.neoforge.client.event.*;
 import org.jetbrains.annotations.NotNull;
 @Mod(value = CreateCyberGoggles.ID, dist = Dist.CLIENT)
 public class CreateCyberGoggles {
 	public static final String ID = "create_cyber_goggles";
-	public static ModConfig config;
 	public CreateCyberGoggles(@NotNull ModContainer container) {
-		config = AutoConfig.register(ModConfig.class, Toml4jConfigSerializer::new).getConfig();
-		container.registerExtensionPoint(
-			IConfigScreenFactory.class,
-			(modContainer, screen) -> AutoConfig.getConfigScreen(ModConfig.class, screen).get()
-		);
+		CyberConfig.register(container);
+	}
+	@EventBusSubscriber(modid = ID, value = Dist.CLIENT, bus = EventBusSubscriber.Bus.GAME)
+	public static class ClientGameEvents {
+		@SubscribeEvent
+		public static void key(InputEvent.Key event) {
+			KeyInput.toggleDiving();
+			KeyInput.openConfigScreen();
+			KeyInput.openStockScreen();
+			KeyInput.previewFilterScreen();
+		}
+		@SubscribeEvent
+		public static void mouseScrollingEvent(InputEvent.MouseScrollingEvent event) {
+			MouseScroll.onMouseScroll(event);
+		}
+	}
+	@EventBusSubscriber(modid = ID, value = Dist.CLIENT, bus = EventBusSubscriber.Bus.MOD)
+	public static class ClientModEvents {
+		@SubscribeEvent
+		public static void registerKeyMappingsEvent(RegisterKeyMappingsEvent event) {
+			ModKeyMapping.register(event);
+		}
+		@SubscribeEvent
+		public static void registerGuiLayersEvent(RegisterGuiLayersEvent event) {
+			OverlayRenderer.register(event);
+		}
 	}
 }