+++ src/main/java/com/forgestove/create_cyber_goggles/mixin/jei/CreateJEIMixin.java
@@ -1,5 +1,5 @@
 package com.forgestove.create_cyber_goggles.mixin.jei;
-import com.forgestove.create_cyber_goggles.CreateCyberGoggles;
+import com.forgestove.create_cyber_goggles.content.config.CyberConfig;
 import com.simibubi.create.compat.jei.CreateJEI;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.*;
@@ -13,7 +13,7 @@ public abstract class CreateJEIMixin {
 	), index = 1
 	)
 	private int loadCategories(int height) {
-		if (!CreateCyberGoggles.config.jei.nonrandomScrap) return height;
+		if (!CyberConfig.get().jei.nonrandomScrap) return height;
 		else return height + 40;
 	}
 }